+++ pkg/eval/builtin_fn_flow_test.go
@@ -23,9 +23,9 @@ func TestEach(t *testing.T) {
 		That(`echo "1\r\n233" | each $put~`).Puts("1", "233"),
 		That(`each $put~ [1 233]`).Puts("1", "233"),
 		That(`range 10 | each [x]{ if (== $x 4) { break }; put $x }`).
-			Puts(int64(0), int64(1), int64(2), int64(3)),
+			Puts(0, 1, 2, 3),
 		That(`range 10 | each [x]{ if (== $x 4) { fail haha }; put $x }`).
-			Puts(int64(0), int64(1), int64(2), int64(3)).Throws(AnyError),
+			Puts(0, 1, 2, 3).Throws(AnyError),
 		// TODO(xiaq): Test that "each" does not close the stdin.
 	)
 }
